refactor AWsKMSWrapper
